refactor(phase4): extract resource_builder from eon_parser.rs
Moves build_enums, build_mixins, expand_resource_mixins,
build_resources_with_enums, and all supporting action/index/field
helpers into eon_parser/resource_builder.rs.

Fixes test compilation by adding explicit imports for model types
and quote::ToTokens into mod tests, and a cfg(test)-scoped
ResourceSpec import for the build_resources helper.

Also cleans up unused imports in resource_struct_tokens.rs
and resource_builder.rs.

Baseline: 233 passing / 19 pre-existing failing (unchanged).
