+++ crates/rest_macro_core/src/compiler/codegen/resource_struct_tokens.rs
@@ -21,8 +21,7 @@ use super::{
     ListBindKind,
     create_payload_field_is_optional, create_payload_field_ty, create_payload_fields,
     enum_values_as_message,
-    hybrid_resource_enforcement,
-    list_bind_ident, list_bind_kinds, list_bind_match_tokens,
+    list_bind_ident, list_bind_kinds,
     list_cursor_payload_ident, list_cursor_value_ident,
     list_filter_field_ty, list_plan_ident,
     list_query_ident, list_response_ident,
